major package move for sharing, to centralize dispersed adaptions
