+++ src/main/java/de/tum/cit/aet/artemis/programming/service/sharing/SharingException.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.artemis.exercise.service.sharing;
+package de.tum.cit.aet.artemis.programming.service.sharing;
 
 import java.io.Serial;
