+++ src/test/java/de/tum/cit/aet/artemis/programming/service/sharing/SharingConnectorServiceTest.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.artemis.exercise.service.sharing;
+package de.tum.cit.aet.artemis.programming.service.sharing;
 
 import static org.assertj.core.api.Assertions.assertThat;
